test (BonusNumberValidatorTest): 보너스 숫자 검증 테스트 이동
