--- src/test/java/lotto/domain/BonusNumberTest.java
@@ -1,5 +1,0 @@
-package lotto.domain;
-
-class BonusNumberTest {
-
-}
